engine: rename snapshot worker to region worker
